Fix some final bugs and add javadocs
