+++ src/main/java/de/bluecolored/bluenbt/BlueNBT.java
@@ -35,6 +35,11 @@ import java.lang.reflect.Type;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 
+/**
+ * The heart of BlueNBT.
+ * <p>Use this class to register your {@link TypeSerializer}s, {@link TypeDeserializer}s and {@link InstanceCreator}s and
+ * (de)serialize any object to/from NBT.</p>
+ */
 public class BlueNBT {
 
     private final List<TypeSerializerFactory> serializerFactories = new ArrayList<>();
@@ -44,11 +49,19 @@ public class BlueNBT {
     private final Map<TypeToken<?>, TypeDeserializer<?>> typeDeserializerMap = new ConcurrentHashMap<>();
     private final Map<TypeToken<?>, InstanceCreator<?>> instanceCreatorMap = new ConcurrentHashMap<>();
 
+    /**
+     * The {@link NamingStrategy} this BlueNBT instance uses to determine the NBT-name
+     * from a given {@link java.lang.reflect.Field}.
+     * <p>Defaults to {@link NamingStrategy#FIELD_NAME}</p>
+     */
     @Getter @Setter
     private NamingStrategy namingStrategy = NamingStrategy.FIELD_NAME;
 
+    /**
+     * Creates a fresh BlueNBT instance with the default configuration and default (de)serializers.
+     */
     public BlueNBT() {
-        register(ObjectAdapterFactory.INSTANCE);
+        register(TypeToken.of(Object.class), ObjectDeserializer.INSTANCE);
         register(ArrayAdapterFactory.INSTANCE);
         register(PrimitiveSerializerFactory.INSTANCE);
         register(PrimitiveDeserializerFactory.INSTANCE);
@@ -56,24 +69,39 @@ public class BlueNBT {
         register(MapAdapterFactory.INSTANCE);
     }
 
+    /**
+     * Registers a {@link TypeAdapterFactory} for this BlueNBT instance to use for (de)serialization.
+     */
     public synchronized void register(TypeAdapterFactory typeAdapterFactory) {
         serializerFactories.add(typeAdapterFactory);
         deserializerFactories.add(typeAdapterFactory);
         typeSerializerMap.clear();
     }
 
+    /**
+     * Registers a {@link TypeSerializerFactory} for this BlueNBT instance to use for serialization.
+     */
     public synchronized void register(TypeSerializerFactory typeSerializerFactory) {
         serializerFactories.add(typeSerializerFactory);
     }
 
+    /**
+     * Registers a {@link TypeDeserializerFactory} for this BlueNBT instance to use for deserialization.
+     */
     public synchronized void register(TypeDeserializerFactory typeDeserializerFactory) {
         deserializerFactories.add(typeDeserializerFactory);
     }
 
+    /**
+     * Registers a {@link InstanceCreatorFactory} for this BlueNBT instance to use for instance-creation.
+     */
     public synchronized void register(InstanceCreatorFactory instanceCreatorFactory) {
         instanceCreatorFactories.add(instanceCreatorFactory);
     }
 
+    /**
+     * Registers a {@link TypeAdapter} for this BlueNBT instance to use for (de)serialization of the specified type.
+     */
     public <T> void register(TypeToken<T> type, TypeAdapter<T> typeAdapter) {
         register(new TypeAdapterFactory() {
             @Override
@@ -86,6 +114,9 @@ public class BlueNBT {
         });
     }
 
+    /**
+     * Registers a {@link TypeSerializer} for this BlueNBT instance to use for serialization of the specified type.
+     */
     public <T> void register(TypeToken<T> type, TypeSerializer<T> typeSerializer) {
         register(new TypeSerializerFactory() {
             @Override
@@ -98,6 +129,9 @@ public class BlueNBT {
         });
     }
 
+    /**
+     * Registers a {@link TypeDeserializer} for this BlueNBT instance to use for deserialization of the specified type.
+     */
     public <T> void register(TypeToken<T> type, TypeDeserializer<T> typeDeserializer) {
         register(new TypeDeserializerFactory() {
             @Override
@@ -110,7 +144,10 @@ public class BlueNBT {
         });
     }
 
-    public <T> void register(Type type, InstanceCreator<T> instanceCreator) {
+    /**
+     * Registers a {@link InstanceCreator} for this BlueNBT instance to use for instance-creation of the specified type.
+     */
+    public <T> void register(TypeToken<T> type, InstanceCreator<T> instanceCreator) {
         register(new InstanceCreatorFactory() {
             @Override
             @SuppressWarnings("unchecked")
@@ -122,6 +159,9 @@ public class BlueNBT {
         });
     }
 
+    /**
+     * Returns the {@link TypeSerializer} for the given type
+     */
     @SuppressWarnings("unchecked")
     public <T> TypeSerializer<T> getTypeSerializer(TypeToken<T> type) {
         TypeSerializer<T> serializer = (TypeSerializer<T>) typeSerializerMap.get(type);
@@ -150,6 +190,9 @@ public class BlueNBT {
         return serializer;
     }
 
+    /**
+     * Returns the {@link TypeDeserializer} for the given type
+     */
     @SuppressWarnings("unchecked")
     public <T> TypeDeserializer<T> getTypeDeserializer(TypeToken<T> type) {
         TypeDeserializer<T> deserializer = (TypeDeserializer<T>) typeDeserializerMap.get(type);
@@ -178,6 +221,9 @@ public class BlueNBT {
         return deserializer;
     }
 
+    /**
+     * Returns the {@link InstanceCreator} for the given type
+     */
     @SuppressWarnings("unchecked")
     public <T> InstanceCreator<T> getInstanceCreator(TypeToken<T> type) {
         InstanceCreator<T> instanceCreator = (InstanceCreator<T>) instanceCreatorMap.get(type);
@@ -206,52 +252,138 @@ public class BlueNBT {
         return instanceCreator;
     }
 
+    /**
+     * Serializes an object to NBT using the specified type for serialization, and writes it to the given
+     * {@link OutputStream}
+     * @param object The Object that should be serialized
+     * @param out The {@link OutputStream} to write the data to
+     * @param type The type that should be used during serialization
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> void write(T object, OutputStream out, TypeToken<T> type) throws IOException {
         write(object, new NBTWriter(out), type);
     }
 
+    /**
+     * Serializes an object to NBT using the specified type for serialization, and writes it to the given
+     * {@link NBTWriter}
+     * @param object The Object that should be serialized
+     * @param out The {@link NBTWriter} to write the data to
+     * @param type The type that should be used during serialization
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> void write(T object, NBTWriter out, TypeToken<T> type) throws IOException {
         getTypeSerializer(type).write(object, out);
     }
 
+    /**
+     * Serializes an object to NBT using the specified {@link Class} for serialization, and writes it to the given
+     * {@link OutputStream}
+     * @param object The Object that should be serialized
+     * @param out The {@link OutputStream} to write the data to
+     * @param type The {@link Class} that should be used during serialization
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> void write(T object, OutputStream out, Class<T> type) throws IOException {
         write(object, out, TypeToken.of(type));
     }
 
+    /**
+     * Serializes an object to NBT using the specified {@link Class} for serialization, and writes it to the given
+     * {@link NBTWriter}
+     * @param object The Object that should be serialized
+     * @param out The {@link NBTWriter} to write the data to
+     * @param type The {@link Class} that should be used during serialization
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> void write(T object, NBTWriter out, Class<T> type) throws IOException {
         write(object, out, TypeToken.of(type));
     }
 
+    /**
+     * Serializes an object to NBT and writes it to the given {@link OutputStream}
+     * @param object The Object that should be serialized
+     * @param out The {@link OutputStream} to write the data to
+     * @throws IOException If an {@link IOException} occurred
+     */
     @SuppressWarnings({"unchecked", "rawtypes"})
     public void write(Object object, OutputStream out) throws IOException {
         write(object, out, (Class) object.getClass());
     }
 
+    /**
+     * Serializes an object to NBT and writes it to the given {@link NBTWriter}
+     * @param object The Object that should be serialized
+     * @param out The {@link NBTWriter} to write the data to
+     * @throws IOException If an {@link IOException} occurred
+     */
     @SuppressWarnings({"unchecked", "rawtypes"})
     public void write(Object object, NBTWriter out) throws IOException {
         write(object, out, (Class) object.getClass());
     }
 
+    /**
+     * Reads an object from the given {@link InputStream} and deserializes it from NBT to the given type.
+     * @param in The {@link InputStream} to read from
+     * @param type The type of the object
+     * @return The deserialized object
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> T read(InputStream in, TypeToken<T> type) throws IOException {
         return read(new NBTReader(in), type);
     }
 
+    /**
+     * Reads an object from the given {@link NBTReader} and deserializes it from NBT to the given type.
+     * @param in The {@link NBTReader} to read from
+     * @param type The type of the object
+     * @return The deserialized object
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> T read(NBTReader in, TypeToken<T> type) throws IOException {
         return getTypeDeserializer(type).read(in);
     }
 
+    /**
+     * Reads an object from the given {@link InputStream} and deserializes it from NBT to the given type.
+     * @param in The {@link InputStream} to read from
+     * @param type The type of the object
+     * @return The deserialized object
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> T read(InputStream in, Class<T> type) throws IOException {
         return read(in, TypeToken.of(type));
     }
 
+    /**
+     * Reads an object from the given {@link NBTReader} and deserializes it from NBT to the given type.
+     * @param in The {@link NBTReader} to read from
+     * @param type The type of the object
+     * @return The deserialized object
+     * @throws IOException If an {@link IOException} occurred
+     */
     public <T> T read(NBTReader in, Class<T> type) throws IOException {
         return read(in, TypeToken.of(type));
     }
 
+    /**
+     * Reads an object from the given {@link InputStream} and deserializes it from NBT to the given type.
+     * @param in The {@link InputStream} to read from
+     * @param type The type of the object
+     * @return The deserialized object
+     * @throws IOException If an {@link IOException} occurred
+     */
     public Object read(InputStream in, Type type) throws IOException {
         return read(in, TypeToken.of(type));
     }
 
+    /**
+     * Reads an object from the given {@link NBTReader} and deserializes it from NBT to the given type.
+     * @param in The {@link NBTReader} to read from
+     * @param type The type of the object
+     * @return The deserialized object
+     * @throws IOException If an {@link IOException} occurred
+     */
     public Object read(NBTReader in, Type type) throws IOException {
         return read(in, TypeToken.of(type));
     }
