+++ src/main/java/de/bluecolored/bluenbt/NBTName.java
@@ -29,14 +29,18 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+/**
+ * Defines one or more fixed nbt-names for the annotated field.<br>
+ * All names of the list will be considered during deserialization, but only the first name will be used for
+ * serialization.
+ * <p>The defined names will ignore any configured {@link NamingStrategy}s.</p>
+ */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.FIELD})
 public @interface NBTName {
 
     /**
-     * Defines one or more nbt-names for the annotated field.<br>
-     * All names of the list will be considered during deserialization, but only the first name will be used for
-     * serialization.
+     * The name(s) that should be used for (de)serialization
      */
     String[] value();
