+++ src/main/java/de/bluecolored/bluenbt/adapter/PrimitiveSerializerFactory.java
@@ -32,6 +32,9 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
+/**
+ * A {@link TypeSerializerFactory} creating {@link TypeSerializer}s for any primitive type and {@link String}s
+ */
 public class PrimitiveSerializerFactory implements TypeSerializerFactory {
 
     public static final PrimitiveSerializerFactory INSTANCE = new PrimitiveSerializerFactory();
@@ -70,11 +73,7 @@ public class PrimitiveSerializerFactory implements TypeSerializerFactory {
     @SuppressWarnings("unchecked")
     public <T> Optional<TypeSerializer<T>> create(TypeToken<T> typeToken, BlueNBT blueNBT) {
         TypeSerializer<?> typeSerializer = TYPE_SERIALIZER_MAP.get(typeToken.getRawType());
-
-        if (typeSerializer != null)
-            return Optional.of((TypeSerializer<T>) typeSerializer);
-
-        return Optional.empty();
+        return Optional.ofNullable((TypeSerializer<T>) typeSerializer);
     }
 
     public static void writeBool(boolean value, NBTWriter writer) throws IOException {
